Added 'COMPARTMENT' as a recognizable node type.
In _process_file(), before generating network in NiceCX from panda dataframe, we replace source and target node names wioth unique IDs of the nodes;
this is done because nodes with the same name are not handled correctly by convert_pandas_to_nice_cx_with_load_plan() as of the moment of writing.
Once network in NiceCX is generated, we replace its noides' names and represents with names.

In _generate_member_node_attributes() changed lisf_of_gene_names_with_prefix to be set set_of_gene_names_with_prefix: this is to ensure
that each symbol is only added once.  Alternatively, we can keep list but check if it already has a given element before adding it to the list.